% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_gpdd_temperature.R
\name{gpdd_location_temperature}
\alias{gpdd_location_temperature}
\title{A function to integrate GPDD time series data with temperature data.}
\usage{
gpdd_location_temperature(location = FALSE, main = TRUE)
}
\arguments{
\item{location}{Do you want to include information on location of the population from the gpdd_location data table? Default is FALSE.
main Do you want to include information on time series from the gpdd_main data table? Default is TRUE.}
}
\description{
This function calls the location data table from rgpdd, and joins it with the temperature data using the LocationID index. 
In addition, this function can be used to join to the GPDD main data table, which contains indexes for all other data tables.
}
\examples{
data <- gpdd_location_temperature() # only include location meta data
head(data)
data_main <- gpdd_location_temperature(location=TRUE, main=TRUE) #include both location meta data and main table.
head(data_main)
}
