% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_gpdd_life_history.R
\name{gpdd_taxon_life_history}
\alias{gpdd_taxon_life_history}
\title{A function to join GPDD time series data with trait data.}
\usage{
gpdd_taxon_life_history(taxon = FALSE, main = TRUE, sources = FALSE)
}
\arguments{
\item{taxon}{Do you want to include taxonomic information from the gpdd_taxon data table? Default is FALSE.}

\item{main}{Do you want to include information on time series from the gpdd_main data table? Default is TRUE.}

\item{sources}{Do you want to include information on data sources for the life history data? Default is FALSE.}
}
\description{
This function calls the taxonomic data table from rgpdd, and joins it with the life history data using the TaxonID index. 
In addition, this function can be used to join to the GPDD main data table, which contains indexes for all other data tables.
}
\examples{
data <- gpdd_taxon_life_history()
head(data)
data_taxon_sources <- gpdd_taxon_life_history(taxon=TRUE, main=FALSE, sources=TRUE)
head(data_taxon_sources)
}
